pyproject.toml
[project]
name = "general-py-discord-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "penne",email = "penne0505pp@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg[binary] (>=3.2,<4.0)",
    "psycopg-pool (>=3.2.7,<4.0.0)",
]

[project.scripts]
general-py-discord-bot = "main:main"

[tool.poetry]
packages = [
    { include = "app" },
    { include = "bot" },
    { include = "main.py" },
]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2,<9.0"
pytest-asyncio = ">=0.25,<0.26"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
